+++ kong/plugins/response-transformer-advanced/constants.lua
@@ -1,0 +1,7 @@
+local constants = {
+  REGEX_SPLIT_RANGE  = "(\\d\\d\\d)-(\\d\\d\\d)",
+  REGEX_SINGLE_STATUS_CODE  = "^\\d\\d\\d$",
+  REGEX_STATUS_CODE_RANGE   = "^\\d\\d\\d-\\d\\d\\d$"
+}
+
+return constants